streamlit_frontend/requirements.txt
streamlit
numpy
Pillow
opencv-python
requests
protobuf
